% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.lda_c}
\alias{as.lda_c}
\title{Converts document-term matrix sparse matrix to 'lda_c' format}
\usage{
as.lda_c(X)
}
\arguments{
\item{X}{Document-Term matrix}
}
\description{
Converts 'dgCMatrix' (or coercible to 'dgCMatrix') to 'lda_c' format
}

