% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analogues.R
\name{check_analogue_accuracy}
\alias{check_analogue_accuracy}
\title{Checks accuracy of word embeddings on analogue task.}
\usage{
check_analogue_accuracy(questions_lst, m_word_vectors, verbose = TRUE)
}
\arguments{
\item{questions_lst}{\code{list} of questions. Each element of \code{questions_lst} is
a \code{integer matrix} with \code{ncol = 4} and represents set of questions
related to particular category. Each element of matrix is a index of row in
\code{m_word_vectors}. See output of \link{prepare_analogue_questions} for details}

\item{m_word_vectors}{word vectors \code{numeric matrix}. Each row should represent word.}

\item{verbose}{\code{logical} whether to print messages during evaluation.}
}
\description{
for full examples see \link{glove}
}
\seealso{
\link{prepare_analogue_questions}, \link{glove}
}

