% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm.R
\name{collect}
\alias{collect}
\alias{collect.RowDistributedMatrix}
\title{Collects distributed data structure to master process}
\usage{
collect(x, ...)

\method{collect}{RowDistributedMatrix}(x, ...)
}
\arguments{
\item{x}{pointer to distributed data structure (only row-distributed matrices at the moment)}

\item{...}{other options (not used at the moment)}
}
\description{
Collects distributed data structure to master process
}
