% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Collocations.R
\docType{data}
\name{Collocations}
\alias{Collocations}
\title{Collocations model.}
\format{\code{\link{R6Class}} object.}
\usage{
Collocations
}
\description{
Creates Collocations model which can be used for phrase extraction.
}
\section{Fields}{

\describe{
\item{\code{collocation_stat}}{\code{data.table} with collocations(phrases) statistics.
Useful for filtering non-relevant phrases}
}}
\section{Usage}{

For usage details see \bold{Methods, Arguments and Examples} sections.
\preformatted{
colloc = Collocations$new(vocabulary, min_collocation_count = 50, sep = "_")
colloc$fit(it)
colloc$transform(it)
colloc$prune(pmi_min = 5, gensim_min = 0, lfmd_min = -Inf)
}
}

\section{Methods}{

\describe{
  \item{\code{$new(vocabulary, min_collocation_count = 50, sep = "_")}}{Constructor for Collocations model.For description of arguments see \bold{Arguments} section.}
  \item{\code{$fit(it)}}{fit Collocations model to input iterator \code{it}}
  \item{\code{$transform(it)}}{transform input iterator using learned collocations
  and produce iterator with collapsed phraeses}
  \item{\code{$transform()}}{ create generator which transforms iterator collapsing phrases}
  \item{\code{$prune(pmi_min = 5, gensim_min = 0, lfmd_min = -Inf)}}{
  filter out non-relevant phrases with low score. User can do it directly by modifying \code{collocation_stat} object.}
}
}

\section{Arguments}{

\describe{
 \item{colloc}{A \code{Collocation} model object}
 \item{it}{An input \code{itoken} iterator}
 \item{vocabulary}{\code{text2vec_vocabulary} - will look for collactions only for words in vocabulary}
}
}
\keyword{datasets}

