% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorizers.R
\name{create_corpus}
\alias{create_corpus}
\title{RAM-friendly streaming corpus construction.}
\usage{
create_corpus(iterator, vectorizer)
}
\arguments{
\item{iterator}{iterator over \code{list} of \code{character} vectors.
Each element is a list of tokens = tokenized and normalized strings.}

\item{vectorizer}{\code{function} vectorizer function.}
}
\value{
corpus object,
We can add documents into this corpus by reference - no copy at all.
See source code for details.
For full process example see \link{get_dtm}.
}
\description{
This functions allow to create corpus objects (vocabulary or hash based),
which are stored outside of R's heap and wrapped via Reference Classes using Rcpp-Modules.
From that objects you can easily extract Document-Term (dtm) and Term-Cooccurnce(tcm)
matrices. Also text2vec grows corpus for \code{tcm} and \code{dtm} simultaneously in a very
ram-friendly and efficient way using iterators abstraction. So you can build corpuses from
objects/files which are orders of magnitude larger that available RAM.
}
\seealso{
\link{vectorizers}
}

