% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distances}
\alias{dist2}
\alias{distances}
\alias{pdist2}
\title{Pairwise Distance Matrix Computation}
\usage{
dist2(x, y = NULL, method = c("cosine", "euclidean", "jaccard"),
  norm = c("none", "l1", "l2"), verbose = TRUE)

pdist2(x, y, method = c("cosine", "euclidean", "jaccard"), norm = c("none",
  "l1", "l2"), verbose = TRUE)
}
\arguments{
\item{x}{first matrix.}

\item{y}{second matrix. \code{y = NULL} is taken to be the same matrix as \code{x}.}

\item{method}{\code{character} or instance of \code{tet2vec_distance} class.
The distance measure to be used. One of \code{c("cosine", "euclidean", "jaccard")} or \link{RWMD}.
\code{RWMD} works only on bag-of-words matrices.
Also user should keep in mind, that distance = 1 - similarity.
\bold{In case of \code{"cosine"} distance function will return \code{1 - abs(similarity)}}}

\item{norm}{\code{character} - how to scale input matrices.}

\item{verbose}{\code{logical} whether to display additional information during calculations}
}
\value{
\code{dist2} returns \code{matrix} of distances between each row of
matrix \code{x} and each row of matrix \code{y}.

\code{pdist2} returns \code{vector} of "parallel" distances between rows
of \code{x} and \code{y}.
}
\description{
\code{dist2} calculates pairwise distances between the
rows of two data matrices. \bold{Note} that some methods work only on sparse matrices and
others work only on dense matrices.

\code{pdist2} calculates "parallel" distances between the rows of two data matrices.
}
\details{
Computes the distance matrix computed by using the specified method.
Similar to \link{dist} function, but works with two matrices.

\code{pdist2} takes two matrices and return a single vector.
giving the ‘parallel’ distances of the vectors.
}

