% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix.R
\name{dtm_get_idf}
\alias{dtm_get_idf}
\title{Inverse Document-Frequency scaling matrix construction}
\usage{
dtm_get_idf(dtm, logScale = log2)
}
\arguments{
\item{logScale}{function to use in idf calculation. Usually \link{log2} used.}

\item{m}{\link{sparseMatrix} - DocumentTermMatrix.}
}
\description{
Creates Inverse Document-Frequency (idf) scaling matrix
from Document-Term-Matrix.
\deqn{idf = \log(\fraq {#documents in the corpus}{# documents where the term + 1}}{%
idf = log (# documents in the corpus) / (# documents where the term appears) }
For examples see  \link{get_dtm}
}

