% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{dtm_get_tf}
\alias{dtm_get_tf}
\title{TermFrequency scaling matrix construction from Document-Term-Matrix}
\usage{
dtm_get_tf(dtm, type = c("tf", "binary"))
}
\arguments{
\item{dtm}{\code{sparseMatrix} - Document-Term-Matrix}

\item{type}{\code{c('tf', 'binary')} - type of TF scaling matrix
Formula for tf :
\code{tf = (Number word appears in document) / (Number words in document)}
\code{tf = Does word appears in document (binary encoding): (0 if not appears), (1 if appears)}}
}
\description{
Creates TermFrequency (tf) scaling matrix from Document-Term-Matrix. For examples
see \link{get_dtm}.
}
\seealso{
\link{dtm_get_idf}, \link{get_dtm}
}

