% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix.R
\name{dtm_remove_common_terms}
\alias{dtm_remove_common_terms}
\title{remove (un)common words (terms) from Document-Term Matrix}
\usage{
dtm_remove_common_terms(dtm, uncommon = 0.01, common = 0.99)
}
\arguments{
\item{m}{\link{sparseMatrix} - TermDocumentMatrix}

\item{sparsity}{maximum level of spasity in resulted matrix}
}
\description{
Creates reduced Document-Term Matrix - throws out uncommon words
to obtain matrix with a given level of sparsity.
}

