% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix.R
\name{dtm_transform}
\alias{dtm_transform}
\title{Transform Document-Term-Matrix to TF/TF-IDF/BOOLEAN form}
\usage{
dtm_transform(dtm, type = c("tfidf", "tf", "boolean"), ...)
}
\arguments{
\item{dtm}{\link{sparseMatrix} - Document-Term-Matrix}

\item{type}{type of transormation. See \link{dtm_get_tf}, \link{dtm_get_idf}}
}
\description{
Creates transformed TF/TF-IDF/BOOLEAN matrix from DocumentTermMatrix.
For example see \link{dtm_get_tf}
}

