% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{filter_commons_transformer}
\alias{filter_commons_transformer}
\title{remove (un)common terms from Document-Term matrix}
\usage{
filter_commons_transformer(dtm, term_freq = c(uncommon = 0.001, common =
  0.975))
}
\arguments{
\item{dtm}{\code{dgCMatrix} - Document-Term Matrix}

\item{term_freq}{- \code{numeric} vector of 2 values in c(0, 1) range
First element corresponds to frequency of uncommon words,
second element corresponds to frequency of common words.
Terms, which are observed less than first value or frequency
or more than second will be filtered out}
}
\description{
Creates reduced Document-Term matrix - throws out
very common and very uncommon words.
}
\seealso{
\link{prune_vocabulary}, \link{tf_transformer},
\link{tfidf_transformer}, \link{binary_transformer}
}

