% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_interface.R
\name{fit}
\alias{fit}
\alias{fit.text2vec_model}
\title{fit model}
\usage{
fit(model, X, ...)

\method{fit}{text2vec_model}(model, X, ...)
}
\arguments{
\item{model}{instance of class \code{text2vec_model}. See \link{lsa}.}

\item{X}{matrix like object. At the moment usually one of
\code{c("matrix", "dgCMatrix", "dgTMatrix", "lda_c")}}

\item{...}{arguments to underlying functions. Currently not used.}
}
\value{
fitted model of class \code{text2vec_model}, which can be used for predictions
via \link{transform.text2vec_model} S3 method.
}
\description{
This is generic function to fit text2vec models (class = "text2vec_model")
}
\examples{
data("movie_review")
N = 100
tokens <- movie_review$review[1:N] \%>\% tolower \%>\% word_tokenizer
dtm <- create_dtm(itoken(tokens), hash_vectorizer())
n_factors = 10
fitted_model = lsa(n_factors) \%>\% fit(dtm)
document_vectors = transform(fitted_model, dtm)
}

