% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_S3.R
\name{fit_transform}
\alias{fit_transform}
\alias{fit_transform.Matrix}
\alias{fit_transform.matrix}
\title{Fit model to data, then transform it}
\usage{
fit_transform(x, model, ...)

\method{fit_transform}{Matrix}(x, model, ...)

\method{fit_transform}{matrix}(x, model, ...)
}
\arguments{
\item{x}{a matrix like object, should inherit from \code{Matrix} or \code{matrix}}

\item{model}{instance of class \code{estimator} which should implement method
with signature \code{$fit(x, ...)}}

\item{...}{additional data/model dependent arguments to downstream functions.
For supervised learning (if any) responese should be called \code{y}.}
}
\value{
Transformed version of \code{x}
}
\description{
This is generic function to fit transformers (class = "transformer")
and then apply fitted model to input.
}

