% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrices.R
\name{get_dtm}
\alias{get_dtm}
\title{DocumentTermMatrix construction}
\usage{
get_dtm(textList, dictionary = NULL)
}
\arguments{
\item{textList}{- \link{character} \link{list} - input text corpus. It should be provided as list of character vectors.
Each element in the list represents document in corpus, each element in character vector represents word(term) in document.}

\item{dictionary}{- \link{character} or \link{NULL} -  use only words from this dictionary in DocumentTermMatrix construction.
NULL if all words should be used.}
}
\description{
Creates DocumentTermMatrix from list of character vectors.
}
\examples{
# create DocumentTermMatrix from two simple texts
textList <- preprocess_corpus(corpus = c("first text", "Second text. It is much much longer!:-)"))
DTM <- get_dtm(textList)
DTMWithDictionary <- get_dtm(textList, dictionary = c("first", "second", "text"))
}

