% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrices.R
\name{get_idf}
\alias{get_idf}
\title{InverseDocumentFrequency scaling matrix construction from DocumentTermMatrix}
\usage{
get_idf(m, logScale = log2, type = "idf")
}
\arguments{
\item{m}{\link{simple_triplet_matrix} - DocumentTermMatrix.}

\item{logScale}{function to use in idf calculation. In usually \link{log2} used.}

\item{type}{type of scaling. Formula for idf:
\deqn{idf = \log(\fraq {#documents in the corpus}{# documents where the term + 1}}{%
idf = log (# documents in the corpus) / (# documents where the term appears) }}
}
\description{
Creates InverseDocumentFrequency (idf) scaling matrix from DocumentTermMatrix
}
\examples{
# create DocumentTermMatrix from two simple texts
textList <- preprocess_corpus(corpus = c("first text", "Second text. It is much much longer!:-)"))
DTM <- get_dtm(textList)
IDFScalingMatrix <- get_idf(DTM)
}

