% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrices.R
\name{get_tdm}
\alias{get_tdm}
\title{TermDocumentMatrix construction}
\usage{
get_tdm(textList, dictionary = NULL, parallerismLevel = 1,
  mc.cores = ceiling(detectCores()/2))
}
\arguments{
\item{textList}{- \link{character} \link{list} - input text corpus. It should be provided as list of character vectors.
Each element in the list represents document in corpus, each element in character vector represents word(term) in document.}

\item{dictionary}{- \link{character} or \link{NULL} -  use only words from this dictionary in TermDocumentMatrix construction.
NULL if all words should be used.}
}
\description{
Creates TermDocumentMatrix from list of character vectors.
}
\examples{
# create TermDocumentMatrix from two simple texts.
textList <- preprocess_corpus(corpus = c("first text", "Second text. It is much much longer!:-)"))
TDM <- get_tdm(textList)
TDMWithDictionary <- get_tdm(textList, dictionary = c("first", "second", "text"))
}

