% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tmlite.R
\name{HashCorpus}
\alias{HashCorpus}
\title{Rcpp module: HashCorpus
Exposes C++ functions to construct hashed Document-Term Matrix}
\description{
Rcpp module: HashCorpus
Exposes C++ functions to construct hashed Document-Term Matrix
}

