% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{ilines}
\alias{ilines}
\title{Creates iterator over lines of connection/file}
\usage{
ilines(con, n, ...)
}
\arguments{
\item{con}{\code{connection} object or a character string.}

\item{n}{\code{integer}, the maximum number of lines to read per iteration.
Negative values indicate that one should read up to the end of the connection.
The default value is 1.}

\item{...}{arguments passed to \code{readLines} function.}
}
\description{
Result of this function usually used in \link{itoken} function.
}
\seealso{
\link{itoken}
}

