% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jsPCA_robust}
\alias{jsPCA_robust}
\title{(numerically robust) Dimension reduction via Jensen-Shannon Divergence & Principal Components}
\usage{
jsPCA_robust(phi)
}
\arguments{
\item{phi}{matrix, with each row containing the distribution over terms
for a topic, with as many rows as there are topics in the model, and as
many columns as there are terms in the vocabulary.}
}
\description{
This function is largely a copy of the repsective function in
https://github.com/cpsievert/LDAvis/blob/master/R/createJSON.R, however,
with a fix to avoid log(0) proposed by Maren-Eckhoff in
https://github.com/cpsievert/LDAvis/issues/56
}
