% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LSA.R
\name{LSA}
\alias{LSA}
\title{Constructor for Latent Semantic Analysis model}
\usage{
LSA(n_topics, verbose = FALSE)
}
\arguments{
\item{n_topics}{number of latent factors
(number of singular values in underlying SVD decomposition).}

\item{verbose}{\code{logical} print status messages}
}
\description{
Wrapper for \code{LatentSemanticAnalysis$new()}
}
\examples{
data("movie_review")
N = 100
tokens <- movie_review$review[1:N] \%>\% tolower \%>\% word_tokenizer
dtm <- create_dtm(itoken(tokens), hash_vectorizer())
lsa_model = LSA(20)
fit(lsa_model, dtm) # same as lsa_model$fit(dtm)
latent_docs = transf(lsa_model, dtm)
}

