% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transformers.R
\name{mutate}
\alias{mutate}
\title{Creates transformed Document-Term matrix.}
\usage{
mutate(dtm, transformer, ...)
}
\arguments{
\item{dtm}{\code{sparseMatrix} - Document-Term-Matrix}

\item{transformer}{- transormation function. Usually one of \link{filter_commons_transformer},
\link{tf_transformer}, \link{tfidf_transformer}, \link{binary_transformer}.}

\item{...}{- transformer parameters}
}
\description{
Transform Document-Term matrix using one of the package-included
transformers or user-specified custom transformer.
For example see \link{dtm_get_tf}
}
\examples{
\dontrun{
txt <- c(paste(letters[c(4:7, 5:12)], collapse = " "),
   paste(LETTERS[c(5:9, 7:12) ], collapse = " "))
corpus <- create_vocab_corpus(txt,
   tokenizer = regexp_tokenizer
   )
# create dtm
dtm <- get_dtm(corpus, dictionary = letters[4:8], stopwords = letters[5:6] ) \%>\%
# filter out very common and very uncommon terms
 mutate(filter_commons_transformer, c(0.001, 0.975))

# simple term-frequency transormation
transformed_tf <- dtm \%>\%
 mutate(tf_transformer)

# tf-idf transormation
transformed_tfidf <- dtm \%>\%
 mutate(tfidf_transformer)
 }
}

