% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ngrams}
\alias{ngrams}
\title{Generates ngrams}
\usage{
ngrams(terms, ngram_min, ngram_max, sep = "_")
}
\arguments{
\item{terms}{input tokens}

\item{ngram_min}{min number of tokens in ngram}

\item{ngram_max}{max number of tokens in ngram}

\item{sep}{string separator between tokems}
}
\description{
Generates ngrams character vector from tokens
}

