% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_interface.R
\name{partial_fit}
\alias{partial_fit}
\alias{partial_fit.text2vec_model}
\title{Fit model to chunk of data}
\usage{
partial_fit(object, X, ...)

\method{partial_fit}{text2vec_model}(object, X, ...)
}
\arguments{
\item{object}{instance of class \code{text2vec_model}.}

\item{X}{matrix like object. At the moment usually one of
\code{c("matrix", "dgCMatrix", "dgTMatrix", "lda_c")}}

\item{...}{arguments to underlying functions. Currently not used.}
}
\value{
Current estimate of model-specific cost function.
}
\description{
This is generic function to fit text2vec models (class = "text2vec_model")
to subset of data. \strong{Note, that this function modifies input model during fitting!}
}

