% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_interface.R
\name{predict}
\alias{predict}
\alias{predict.text2vec_model}
\title{Transform new data with fitted model}
\usage{
\method{predict}{text2vec_model}(object, X, ...)
}
\arguments{
\item{object}{instance of class \code{text2vec_model}.}

\item{X}{matrix like object. At the moment usually one of
\code{c("matrix", "dgCMatrix", "dgTMatrix", "lda_c")}}

\item{...}{arguments to underlying functions. Currently not used.}
}
\value{
Transformed version of \code{X}
}
\description{
This is generic function to apply fitted text2vec models
(class = "text2vec_model") to new data.
}
\examples{
data("movie_review")
N = 100
tokens <- movie_review$review[1:N] \%>\% tolower \%>\% word_tokenizer
dtm <- create_dtm(itoken(tokens), hash_vectorizer())
n_factors = 10
fitted_model = LSA(n_factors) \%>\% fit(dtm)
document_vectors = predict(fitted_model, dtm)
}

