% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analogues.R
\name{prepare_analogue_questions}
\alias{prepare_analogue_questions}
\title{Prepares questions list from \code{questions-words.txt} format.}
\usage{
prepare_analogue_questions(questions_file_path, vocab_terms, verbose = TRUE)
}
\arguments{
\item{questions_file_path}{\code{character} path of questions file.}

\item{vocab_terms}{\code{character vector}, words which we have in vocabulary
and word embeddings matrix.}

\item{verbose}{\code{logical} whether to print messages during evaluation.}
}
\description{
for full examples see \link{glove}
}
\seealso{
\link{check_analogue_accuracy}, \link{glove}
}

