% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analogies.R
\name{prepare_analogy_questions}
\alias{prepare_analogy_questions}
\title{Prepares list of analogy questions}
\usage{
prepare_analogy_questions(questions_file_path, vocab_terms, verbose = TRUE)
}
\arguments{
\item{questions_file_path}{\code{character} path to questions file.}

\item{vocab_terms}{\code{character} words which we have in the
vocabulary and word embeddings matrix.}

\item{verbose}{\code{logical} whether to print messages during evaluation.}
}
\description{
This function prepares a list of questions from a
  \code{questions-words.txt} format. For full examples see \link{GloVe}.
}
\seealso{
\link{check_analogy_accuracy}, \link{GloVe}
}

