% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{print.text2vec_vocabulary}
\alias{print.text2vec_vocabulary}
\title{Printing Vocabulary}
\usage{
\method{print}{text2vec_vocabulary}(x, ...)
}
\arguments{
\item{x}{vocabulary}

\item{...}{optional arguments to print methods.}
}
\description{
Print a vocabulary.
}
