% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{prune_vocabulary}
\alias{prune_vocabulary}
\title{Prunes vocabulary.}
\usage{
prune_vocabulary(vocabulary, term_count_min = 1L, term_count_max = Inf,
  doc_proportion_min = 0, doc_proportion_max = 1,
  max_number_of_terms = Inf)
}
\arguments{
\item{vocabulary}{vocabulary from \link{vocabulary} function output.}

\item{term_count_min}{minimum number of occurences over all documents.}

\item{term_count_max}{maximum number of occurences over all documents.}

\item{doc_proportion_min}{minimum proportion of documents which should contain term.}

\item{doc_proportion_max}{maximum proportion of documents which should contain term.}

\item{max_number_of_terms}{maximum number of terms in vocabulary.}
}
\description{
Perform filtering of the input vocabulary and trhrowing out very frequet and
very infrequet terms. Also leaves top `max_number_of_terms` (by count) terms. See examples
in \link{vocabulary} function.
}
\seealso{
\link{vocabulary}
}

