% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2vec.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fit}
\alias{fit_transform}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{mlapi}{\code{\link[mlapi]{fit}}, \code{\link[mlapi]{fit_transform}}}
}}

