% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrices.R
\name{remove_sparse_terms}
\alias{remove_sparse_terms}
\title{remove sparse words (terms) from TermDocumentMatrix matrix}
\usage{
remove_sparse_terms(m, sparsity)
}
\arguments{
\item{m}{\link{simple_triplet_matrix} - TermDocumentMatrix}

\item{sparsity}{maximum level of spasity in resulted matrix}
}
\description{
Creates reduced TermDocumentMatrix - throws out uncommon words
to obtain matrix with a given level of sparsity.
}
\examples{
# create DocumentTermMatrix from two simple texts
textList <- preprocess_corpus(corpus = c("first text", "Second text. It is much much longer!:-)"))
DTM <- get_dtm(textList)
DTMTransformed <- transform_dtm(DTM, type='tfidf')
}

