% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_RWMD.R
\name{RWMD}
\alias{RWMD}
\title{Creates model which can be used for calculation of "relaxed word movers distance".}
\usage{
RWMD(word_vectors, method = c("cosine", "euclidean"), normalize = TRUE)
}
\arguments{
\item{word_vectors}{numeric matrix which contains word embeddings. Rows - words,
columns - corresponding vectors. Rows should have names (words).}

\item{method}{name of the distance for measuring similarity between word vectors.
\code{"cosine"} by default. This means distance = 1 - cosine angle betwen vectors.}

\item{normalize}{should function perform L2 normalization of word vectors?
Word vectors should be normalized before calculation of distances between them!
If user has pre-normalized input, he/she can set \code{normalize = FALSE}.}
}
\description{
Relaxed word movers distance tries to measure distance between documents by
calculating how hard is to transofrm words from first document into words from second document
and vice versa. For more detail see original article: \url{http://mkusner.github.io/publications/WMD.pdf}.
}
\examples{
\dontrun{
data("movie_review")
tokens = movie_review$review \%>\%
  tolower \%>\%
  word_tokenizer
v = create_vocabulary(itoken(tokens)) \%>\%
  prune_vocabulary(term_count_min = 5, doc_proportion_max = 0.5)
corpus = create_corpus(itoken(tokens), vocab_vectorizer(v, skip_grams_window = 5))
dtm = get_dtm(corpus)
tcm = get_tcm(corpus)
glove_model = GloVe(word_vectors_size = 50, vocabulary = v, x_max = 10)
wv = fit_predict(glove_model, tcm, n_iter = 10)
rwmd_model = RWMD(word_vectors = wv)
rwmd_dist = dist2(dtm[1:10, ], dtm[1:100, ], metrics = rwmd_model, norm = 'none')
}
}

