% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{simple_preprocess}
\alias{simple_preprocess}
\title{simple standart text preprocessing function.}
\usage{
simple_preprocess(txt, change_case_func = tolower, clean_func = (function(x)
  gsub(pattern = "[^a-z]", replacement = " ", x = x)), strip_space = T,
  trim_space = T)
}
\arguments{
\item{change_case_func}{- \link{function} performing case transformation -
convert text between upper case, lower case. \link{tolower} by default.
It should take \code{caracter} vectors and return \code{caracter} vectors.
Use \link{identity} or \code{NULL} if no transformation needed.}

\item{clean_func}{- \link{function} performing text cleaning.
By default it only keep latin symbols '[a-z]+' and replace other symbols by
whitespace. Should take \code{caracter} vectors and return \code{caracter} vectors.
Use \link{identity} or \code{NULL} if no transformation needed.}

\item{strip_space}{- \link{logical}, wheteher replace multiple whitespace by signle one.}

\item{trim_space}{- \link{logical}, wheteher remove trailing and leading space.}
}
\value{
- \link{character} vectors - preprocessed text.
}
\description{
simple standart text preprocessing function.
}
\examples{
doc <- c(" first  second", "bla, bla, blaa 33  ")
simple_preprocess(doc)
}

