% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_into}
\alias{split_into}
\title{split vector into N roughly equal parts}
\usage{
split_into(vec, N)
}
\arguments{
\item{vec}{unput \code{vector}}

\item{N}{\code{integer} desired number of chunks}
}
\value{
\link{list} with \code{N} roughly equal length elements
}
\description{
splits vector into N roughly equal parts
}

