% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2vec.R
\docType{package}
\name{text2vec}
\alias{text2vec}
\alias{text2vec-package}
\title{text2vec is a package that provides an efficient framework with a concise API
for text analysis and natural language processing in R.}
\description{
text2vec is a package that provides an efficient framework with a concise API
for text analysis and natural language processing in R.
}

