% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_lda_c}
\alias{to_lda_c}
\title{Converts 'dgCMatrix' to 'lda_c' format}
\usage{
to_lda_c(dtm)
}
\arguments{
\item{dtm}{Document-Term matrix}
}
\description{
Converts 'dgCMatrix' (or coercible to 'dgCMatrix') to 'lda_c' format
}

