% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transformers.R
\name{transform}
\alias{transform}
\alias{transform.Matrix}
\title{Creates transformed Document-Term matrix.}
\usage{
\method{transform}{Matrix}(dtm, transformer, ...)
}
\arguments{
\item{dtm}{\link{sparseMatrix} - Document-Term-Matrix}

\item{transformer}{- transormation function. Usually one of \link{filter_commons_transformer},
\link{tf_transformer}, \link{idf_transformer}, \link{binary_transformer}.}

\item{...}{- transformer parameters}
}
\description{
Transform Document-Term matrix using one of the package-included
transformers or user-specified custom transformer.
For example see \link{dtm_get_tf}
}
\examples{
txt <- c(paste(letters[c(4:7, 5:12)], collapse = " "), paste(LETTERS[c(5:9, 7:12) ], collapse = " "))
corpus <- create_dict_corpus(txt,
   tokenizer = simple_tokenizer
   )
# create dtm
dtm <- get_dtm(corpus, dictionary = letters[4:8], stopwords = letters[5:6] ) \%>\%
# filter out very common and very uncommon terms
 transform(filter_commons_transformer, c(0.001, 0.975))

# simple term-frequency transormation
transformed_tf <- dtm \%>\%
 transform(tf_transformer)

# tf-idf transormation
transformed_tfidf <- dtm \%>\%
 transform(tfidf_transformer)
}

