% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrices.R
\name{transform_dtm}
\alias{transform_dtm}
\title{Transform DocumentTermMatrix of wordcounts to TF/TF-IDF/BOOLEAN form}
\usage{
transform_dtm(m, type = c("tfidf", "tf", "boolean"), ...)
}
\arguments{
\item{m}{\link{simple_triplet_matrix} - DocumentTermMatrix.}

\item{type}{type of transormation. See \link{get_tf}, \link{get_idf}}
}
\description{
Creates transformed TF/TF-IDF/BOOLEAN matrix from DocumentTermMatrix of wordcounts
}
\examples{
# create DocumentTermMatrix from two simple texts
textList <- preprocess_corpus(corpus = c("first text", "Second text. It is much much longer!:-)"))
DTM <- get_dtm(textList)
DTMTransformed <- transform_dtm(DTM, type='tfidf')
}

