% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{transform_filter_commons}
\alias{transform_filter_commons}
\title{Remove terms from a document-term matrix}
\usage{
transform_filter_commons(dtm, term_freq = c(uncommon = 0.001, common = 0.975))
}
\arguments{
\item{dtm}{a document-term matrix of class \code{dgCMatrix} or
\code{dgTMatrix}.}

\item{term_freq}{\code{numeric} vector of 2 values in between \code{0} and
\code{1}. The first element corresponds to frequency of uncommon words; the
second element corresponds to the frequency of common words. Terms which
are observed less than first value or frequency or more than second will be
filtered out.}
}
\description{
This function removes very common and very uncommon words from a
  document-term matrix.
}
\seealso{
\link{prune_vocabulary}, \link{transform_tf},
  \link{transform_tfidf}, \link{transform_binary}
}

