% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2vec.R
\name{VocabCorpus}
\alias{VocabCorpus}
\title{Rcpp module: VocabCorpus
Exposes C++ functions to construct Document-Term Matrix}
\description{
Rcpp module: VocabCorpus
Exposes C++ functions to construct Document-Term Matrix
}

