% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2vec.R
\name{VocabularyBuilder}
\alias{VocabularyBuilder}
\title{Rcpp module: VocabularyBuilder
Exposes C++ functions to construct Vocabulary}
\description{
Rcpp module: VocabularyBuilder
Exposes C++ functions to construct Vocabulary
}

