% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm.R
\name{create_dtm}
\alias{create_dtm}
\alias{create_dtm.itoken}
\alias{create_dtm.list}
\title{Document-Term Matrix construction.}
\usage{
create_dtm(itoken_src, vectorizer, type = c("dgCMatrix", "dgTMatrix",
  "lda_c"), ...)

\method{create_dtm}{itoken}(itoken_src, vectorizer, type = c("dgCMatrix",
  "dgTMatrix", "lda_c"), ...)

\method{create_dtm}{list}(itoken_src, vectorizer, type = c("dgCMatrix",
  "dgTMatrix", "lda_c"), verbose = FALSE, ...)
}
\arguments{
\item{itoken_src}{\code{list} of iterators over tokens - \code{itoken}.
Each element is a list of tokens = tokenized and normalized strings.}

\item{vectorizer}{\code{function} vectorizer function.}

\item{type}{character, one of \code{c("dgCMatrix", "dgTMatrix", "lda_c")}.
"lda_c" - Blei's lda-c format (list of 2 * doc_terms_size),
see \url{https://www.cs.princeton.edu/~blei/lda-c/readme.txt}}

\item{...}{- arguments to \link{foreach} function which is used to iterate
over \code{itoken_src} under the hood.}

\item{verbose}{\code{logical} print status messages}
}
\value{
Document-Term Matrix
}
\description{
High-level function for Document-Term Matrix construction.
If parallel backend is registered, it will construct DTM in multiple threads.
}
\details{
User should keep in mind, that he/she should split data itself and
provide list of \link{itoken} iterators. Each element of \code{itoken_src}
will be handled in separate thread and at the end they will be combined.
}
\examples{
\dontrun{
data("movie_review")
N <- 1000
it <- itoken(movie_review$review[1:N], preprocess_function = tolower,
             tokenizer = word_tokenizer)
v <- create_vocabulary(it)
#remove very common and uncommon words
pruned_vocab = prune_vocabulary(v, term_count_min = 10,
 doc_proportion_max = 0.5, doc_proportion_min = 0.001)
vectorizer <- vocab_vectorizer(v)
it <- itoken(movie_review$review[1:N], preprocess_function = tolower,
             tokenizer = word_tokenizer)
dtm <- create_dtm(it, vectorizer)
# get tf-idf matrix from bag-of-words matrix
dtm_tfidf <- transformer_tfidf(dtm)

## Example of parallel mode
# set to number of cores on your machine
N_WORKERS <- 1
doParallel::registerDoParallel(N_WORKERS)
splits <- split_into(movie_review$review, N_WORKERS)
jobs <- lapply(splits, itoken, tolower, word_tokenizer, chunks_number = 1)
vectorizer <- hash_vectorizer()
dtm <- create_dtm(jobs, vectorizer, type = 'dgTMatrix')
}
}
\seealso{
\link{itoken} \link{vectorizers} \link{create_corpus} \link{get_dtm}
}

