% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tmlite.R
\name{DictCorpus}
\alias{DictCorpus}
\title{Rcpp module: DictCorpus
Exposes C++ functions to construct Document-Term Matrix}
\description{
Rcpp module: DictCorpus
Exposes C++ functions to construct Document-Term Matrix
}

