% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix.R
\name{dtm_get_idf}
\alias{dtm_get_idf}
\title{Inverse Document-Frequency scaling matrix construction}
\usage{
dtm_get_idf(dtm, logScale = log2)
}
\arguments{
\item{dtm}{\link{dgCMatrix} - Document-Term matrix.}

\item{logScale}{function to use in idf calculation. Usually \link{log2} used.}
}
\description{
Creates Inverse Document-Frequency (idf) scaling matrix from Document-Term matrix.
idf = log (# documents in the corpus) / (# documents where the term appears + 1)
For examples see  \link{get_dtm}
}

