% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{dtm_get_tf}
\alias{dtm_get_tf}
\title{TermFrequency scaling matrix construction from Document-Term-Matrix}
\usage{
dtm_get_tf(dtm, norm = c("l1", "l2"))
}
\arguments{
\item{dtm}{\code{sparseMatrix} - Document-Term-Matrix}

\item{norm}{\code{character} - Norm used to normalize term vectors. 'l1' by default, i.e.
scale by bumber of words in document.}
}
\description{
Creates TermFrequency (tf) scaling matrix from Document-Term-Matrix. For examples
see \link{get_dtm}.
}
\seealso{
\link{dtm_get_idf}, \link{get_dtm}
}

