% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_GloVe.R, R/models_interface.R
\name{fit.GloVe}
\alias{fit}
\alias{fit.GloVe}
\alias{fit.text2vec_model}
\title{fit text2vec model}
\usage{
\method{fit}{GloVe}(object, X, n_iter, convergence_tol = -1, ...)

fit(object, X, ...)

\method{fit}{text2vec_model}(object, X, ...)
}
\arguments{
\item{object}{instance of class \code{text2vec_model}. See \link{LSA}.}

\item{X}{matrix like object. At the moment usually one of
\code{c("matrix", "dgCMatrix", "dgTMatrix", "lda_c")}}

\item{n_iter}{number of iterations}

\item{convergence_tol}{convergence tolerance}

\item{...}{arguments to underlying functions. Currently not used.}
}
\value{
fitted object of class \code{text2vec_model}, which can be used for transformations
via \link{transf.text2vec_model} S3 method.
}
\description{
This is generic function to fit text2vec models (class = "text2vec_model")
}
\examples{
data("movie_review")
N = 100
tokens <- movie_review$review[1:N] \%>\% tolower \%>\% word_tokenizer
dtm <- create_dtm(itoken(tokens), hash_vectorizer())
n_factors = 10
fitted_model = LSA(n_factors) \%>\% fit(dtm)
document_vectors = transf(fitted_model, dtm)
}

