% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_S3.R
\name{fit}
\alias{fit}
\alias{fit.Matrix}
\alias{fit.matrix}
\title{Fits model to data}
\usage{
fit(x, model, y = NULL, ...)

\method{fit}{Matrix}(x, model, y = NULL, ...)

\method{fit}{matrix}(x, model, y = NULL, ...)
}
\arguments{
\item{x}{a matrix like object, should inherit from \code{Matrix} or \code{matrix}}

\item{model}{instance of class \code{estimator} which should implement method
with signature \code{$fit(x, y, ...)}}

\item{y}{\code{NULL} by default. Optional response variable for supervised models.
Should inherit from \code{vector} \code{Matrix} or \code{matrix}. See documentation
for corresponding models.}

\item{...}{additional data/model dependent arguments to downstream functions.}
}
\value{
\code{invisible(object$self())}
}
\description{
Generic function to fit models - inherited from \code{estimator}
}

