% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix.R
\name{get_dtm}
\alias{get_dtm}
\alias{get_tdm}
\title{DTM ( TDM ) construction}
\usage{
get_dtm(corpus, dictionary = NULL, stopwords = NULL, type = c("dgCMatrix",
  "dgTMatrix", "LDA_C"))

get_tdm(corpus, dictionary = NULL, stopwords = NULL)
}
\arguments{
\item{corpus}{- TmliteCorpus input corpus.
It can be obtained via \link{create_dict_corpus} function}

\item{dictionary}{- \link{character} or \link{NULL} -  use only words from this dict
dictionary in Document-Term-Matrix construction.
NULL if all words should be used.}

\item{stopwords}{- \link{character} or \link{NULL} - words to remove from DTM ( TDM )}
}
\description{
Creates Document-Term-Matrix/Term-Document-Matrix from TmliteCorpus.
}
\section{Functions}{
\itemize{
\item \code{get_tdm}: 
}}
\examples{
preprocess_fun <- function(txt) {
   txt \%>\%
     tolower \%>\%
     # keep only words with latin letters
     gsub(pattern = "[^a-z]", replacement = " ", x = .) \%>\%
     # strip whitespaces
     gsub(pattern = "\\\\s+", replacement = " ", x = .)
}
stemfun <- function(txt_char_vec, lang = 'en')
 lapply(txt_char_vec, function(x) SnowballC::wordStem(x, language = lang))
# or use simple_preprocess() insted
txt <- c(paste(letters[c(4:7, 5:12)], collapse = " "), paste(LETTERS[c(5:9, 7:12) ], collapse = " "))
corpus <- create_dict_corpus(txt,
   preprocess_fun = preprocess_fun,
   tokenizer = simple_tokenizer,
   stemming_fun = stemfun,
   batch_size = 1
   )
dtm <- get_dtm(corpus, dictionary = letters[4:8], stopwords = letters[5:6] )
tdm <- get_tdm(corpus, dictionary = letters[4:8], stopwords = letters[5:6] )
}

