% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm.R
\name{get_dtm}
\alias{get_dtm}
\title{Creates Document-Term matrix}
\usage{
get_dtm(corpus, type = c("dgCMatrix", "dgTMatrix", "lda_c"))
}
\arguments{
\item{corpus}{HashCorpus or VocabCorpus object.
See \link{create_corpus} for details.}

\item{type}{character, one of \code{c("dgCMatrix", "dgTMatrix", "lda_c")}.
"lda_c" - Blei's lda-c format (list of 2*doc_terms_size),
see \url{https://www.cs.princeton.edu/~blei/lda-c/readme.txt}}
}
\description{
Creates Document-Term matrix from Corpus object.
}
\examples{
\dontrun{
N <- 1000
it <- itoken(movie_review$review[1:N], preprocess_function = tolower,
             tokenizer = word_tokenizer, chunks_number = 10)
v <- vocabulary(it, c(1L, 1L) )

#remove very common and uncommon words
pruned_vocab = prune_vocabulary(v, term_count_min = 10,
 doc_proportion_max = 0.8, doc_proportion_min = 0.001,
 max_number_of_terms = 10000)

vectorizer <- vocab_vectorizer(v)
it <- itoken(movie_review$review[1:N], preprocess_function = tolower,
             tokenizer = word_tokenizer, chunks_number = 10)
corpus <- create_corpus(it, vectorizer)
dtm <- get_dtm(corpus)

tf_scale_matrix <- get_tf(dtm, type = 'tf')
dtm_tf <- tf_scale_matrix \%*\% dtm
dtm_tf_idf <- get_tf \%*\% m \%*\% get_idf(dtm)

# The same result we can obtain using tfidf_transformer function
dtm_tf_idf_2 <- tfidf_transformer(dtm)
identical(dtm_tf_idf, dtm_tf_idf_2)
}
}

