% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{get_tf}
\alias{get_tf}
\title{Term-frequency scaling matrix}
\usage{
get_tf(dtm, norm = c("l1", "l2"))
}
\arguments{
\item{dtm}{a document-term matrix of class \code{dgCMatrix} or
\code{dgTMatrix}.}

\item{norm}{\code{character} the method used to normalize term vectors.
\code{"l1"} by default, i.e., scale by the number of words in the document.}
}
\value{
\code{ddiMatrix} TF scaling diagonal sparse matrix.
}
\description{
This function creates a term-frequency (TF) scaling matrix from
  a document-term matrix.
}
\seealso{
\link{get_idf}, \link{get_dtm}, \link{create_dtm}
}

