% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2vec.R
\docType{data}
\name{GloVe}
\alias{GloVe}
\alias{GlobalVectors}
\title{re-export rsparse::GloVe}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
GlobalVectors
}
\description{
re-export rsparse::GloVe
}
\keyword{datasets}
