% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text2vec.R
\name{GloveCorpus}
\alias{GloveCorpus}
\title{Rcpp module: GloveCorpus
Exposes C++ functions to construct word cooccurence matrix
and train GloVe model for word embeddings}
\description{
Rcpp module: GloveCorpus
Exposes C++ functions to construct word cooccurence matrix
and train GloVe model for word embeddings
}

