% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2vec.R
\name{HashCorpus}
\alias{HashCorpus}
\title{Rcpp module: HashCorpus
Exposes C++ functions to construct hashed Document-Term Matrix}
\description{
Rcpp module: HashCorpus
Exposes C++ functions to construct hashed Document-Term Matrix
}

