% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{ifiles}
\alias{idir}
\alias{ifiles}
\title{Creates iterator over text files from the disk}
\usage{
ifiles(file_paths, reader_function = readLines, ...)

idir(path, reader_function = readLines, ...)
}
\arguments{
\item{file_paths}{\code{character} paths of input files}

\item{reader_function}{\code{function} which will perform reading of text
files from disk, which should take a path as its first argument.}

\item{...}{arguments passed to other methods (including
\code{reader_function}).}

\item{path}{\code{character} path of directory. All files in the directory will be read.}
}
\description{
The result of this function usually used in an \link{itoken}
  function.
}
\examples{
current_dir_files <- list.files(path = ".", full.names = TRUE)
files_iterator <- ifiles(current_dir_files)
dir_files_iterator <- idir(path = ".")
}
\seealso{
\link{itoken}
}

