% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gibbs_LDA.R
\name{LDA}
\alias{LDA}
\title{Creates Latent Dirichlet Allocation model.}
\usage{
LDA(n_topics, vocabulary, doc_topic_prior = 1/n_topics,
  topic_word_prior = 1/n_topics, ...)
}
\arguments{
\item{n_topics}{desired number of topics. Also knows as \bold{K}.}

\item{vocabulary}{vocabulary in a form of \code{character} vector or class of
\code{text2vec_vocab}}

\item{doc_topic_prior}{prior for document-topic multinomial distribution.
Also knows as \bold{alpha}.}

\item{topic_word_prior}{prior for topic-word multinomial distribution.
Also knows as \bold{eta}.}

\item{...}{arguments passed to other methods (not used at the moment).}
}
\description{
\bold{Iterative algorithm}. Model can be fitted
via Collapsed Gibbs Sampling algorithm using \code{fit} or \code{fit_predict} methods.
}

