% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_LSA.R
\name{LSA}
\alias{LSA}
\title{Creates LSA model}
\usage{
LSA(n_topics, verbose = FALSE)
}
\arguments{
\item{n_topics}{number of latent factors
(number of singular values in underlying SVD decomposition).}

\item{verbose}{\code{logical} print status messages}
}
\description{
This function creates LSA(Latent semantic analysis) model.
See \url{https://en.wikipedia.org/wiki/Latent_semantic_analysis} for details.
Model can be fitted with \link{fit}, \link{fit_transform} methods.
}
\examples{
data("movie_review")
N = 100
tokens <- movie_review$review[1:N] \%>\% tolower \%>\% word_tokenizer
dtm <- create_dtm(itoken(tokens), hash_vectorizer())
n_topics = 10
fitted_model_1 = LSA(n_topics) \%>\% fit(dtm)
lsa_model_2 = LSA(n_topics)

all.equal(fit_transform(lsa_model_2, dtm), transform(fitted_model_1, dtm))
}

