% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{mutate}
\alias{mutate}
\title{Creates transformed Document-Term matrix.}
\usage{
mutate(dtm, transformer, ...)
}
\arguments{
\item{dtm}{\code{sparseMatrix} - Document-Term-Matrix}

\item{transformer}{- transormation function. Usually one of
\link{tf_transformer}, \link{tfidf_transformer},
\link{binary_transformer}, \link{filter_commons_transformer}.}

\item{...}{- transformer parameters}
}
\description{
Transform Document-Term matrix using one of the package-included
transformers or user-specified custom transformer.
For example see \link{dtm_get_tf}
}
\examples{
\dontrun{
data(moview_review)

txt <- movie_review[['review']][1:1000]
it <- itoken(txt, tolower, regexp_tokenizer)
vocab <- vocabulary(it)
#remove very common and uncommon words
pruned_vocab = prune_vocabulary(vocab, term_count_min = 10,
 doc_proportion_max = 0.8, doc_proportion_min = 0.001, max_number_of_terms = 20000)

it <- itoken(txt, tolower, regexp_tokenizer)
corpus <- create_vocab_corpus(it, pruned_vocab)
dtm <- get_dtm(corpus, type = 'dgCMatrix' )

dtm_filtered <- dtm \%>\%
 # filter out very common and very uncommon terms
 mutate(filter_commons_transformer, c(0.001, 0.975))

# simple term-frequency transormation
transformed_tf <- dtm \%>\%
 mutate(tf_transformer)

# tf-idf transormation
idf <- dtm_get_idf(dtm)
transformed_tfidf <- dtm \%>\%
 mutate(tfidf_transformer, idf)
 }
}

