% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2vec.R, R/tokenizers.R
\docType{import}
\name{reexports}
\alias{\%>\%}
\alias{boundary}
\alias{coll}
\alias{fixed}
\alias{reexports}
\alias{regex}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{stringr}{\code{\link[stringr]{fixed}}, \code{\link[stringr]{coll}}, \code{\link[stringr]{regex}}, \code{\link[stringr]{boundary}}}
}}
\keyword{internal}

