% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_into}
\alias{split_into}
\title{Split a vector for parallel processing}
\usage{
split_into(vec, n)
}
\arguments{
\item{vec}{input vector}

\item{n}{\code{integer} desired number of chunks}
}
\value{
\code{list} with \code{n} elements, each of roughly equal length
}
\description{
This function splits a vector into \code{n} parts of roughly
  equal size. These splits can be used for parallel processing. In general,
  \code{n} should be equal to the number of jobs you want to run, which
  should be the number of cores you want to use.
}
