% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2vec.R
\docType{package}
\name{text2vec}
\alias{text2vec}
\alias{text2vec-package}
\title{text2vec}
\description{
Fast vectorization, topic modeling, distances and GloVe word embeddings in R.
}
\details{
To learn more about text2vec visit project website: \url{http://text2vec.org}
Or start with the vignettes:
\code{browseVignettes(package = "text2vec")}
}
