% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transformers.R
\name{tf_transformer}
\alias{binary_transformer}
\alias{tf_transformer}
\alias{tfidf_transformer}
\title{Scales Document-Term matrix}
\usage{
tf_transformer(dtm)

tfidf_transformer(dtm, idf = NULL)

binary_transformer(dtm)
}
\arguments{
\item{dtm}{\code{dgCMatrix} - Document-Term matrix}

\item{idf}{- \code{ddiMatrix} \code{Diagonal} matrix for idf-scaling. See \link{dtm_get_idf}.
If not provided ( \code{NULL} ) - idf will be calculated form current data.}
}
\description{
\code{tf_transformer} scales each document vector by # of terms in corresponding document.

\code{tf = (Number word appears in document) / (Number words in document) }

\code{binary_transformer} store 1 if document contains term and 0 otherwise.


\code{tfidf_transformer}

\code{idf  = log (Number documents in the corpus) / (Number documents where the term appears + 1)}
}
\section{Functions}{
\itemize{
\item \code{tfidf_transformer}: 

\item \code{binary_transformer}: 
}}
\seealso{
\link{dtm_get_idf}, examples provided in \link{mutate}
}

