% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transformers.R
\name{tf_transformer}
\alias{binary_transformer}
\alias{tf_transformer}
\alias{tfidf_transformer}
\title{Scales Document-Term matrix}
\usage{
tf_transformer(dtm)

tfidf_transformer(dtm, idf = NULL)

binary_transformer(dtm)
}
\arguments{
\item{dtm}{\link{dgCMatrix} - Document-Term matrix}

\item{idf}{- \link{ddiMatrix} \link{Diagonal} matrix for idf-scaling. See \link{dtm_get_idf}.
If not provided ( \code{NULL} ) - idf will be calculated form current data.}
}
\description{
\code{tf_transformer} scales each document vector by # of terms in corresponding document.

\deqn{tf = \fraq {# word appears in document}{# words in document}}{%
tf = (# word appears in document) / (# words in document) }

\code{binary_transformer} store 1 if document contains term and 0 otherwise.

\deqn{binary = {Does word appears in document (binary encoding): 0 if not appears, 1 if appears}}{%
tf = (Does word appears in document (binary encoding): 0 if not appears, 1 if appears)}

\code{tfidf_transformer}

\deqn{idf = {log (# documents in the corpus) / (# documents where the term appears + 1)}}{%
idf  = log (# documents in the corpus) / (# documents where the term appears + 1)}
}
\section{Functions}{
\itemize{
\item \code{tfidf_transformer}: 

\item \code{binary_transformer}: 
}}
\seealso{
\link{dtm_get_idf}, examples provided in \link{transform}
}

